Outsource helper functions and classes used by FormBackend to separate module formutils
